+++ package.json
@@ -21,7 +21,8 @@
     "react-scripts": "4.0.2",
     "styled-components": "^5.2.1",
     "typescript": "^4.1.5",
-    "web-vitals": "^1.1.0"
+    "web-vitals": "^1.1.0",
+    "yarn": "^1.22.10"
   },
   "scripts": {
     "start": "react-scripts start",
