--- src/components/Description/Description.tsx
@@ -1,10 +1,0 @@
-import { FC } from 'react';
-import StyledDescription from './Description.styles';
-
-interface DescriptionProps {
-  description: string;
-}
-
-const Description: FC<DescriptionProps> = ({ description }) => <StyledDescription>{description}</StyledDescription>;
-
-export default Description;
