--- src/components/Layout/Header/Header.styles.tsx
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-export const HeaderWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  position: fixed;
-  top: 50px;
-  width: 80%;
-`;
