--- src/components/Layout/Layout.styles.tsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const PagesWrapper = styled.div`
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-`;
