+++ src/components/organisms/Header/Header.styles.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const HeaderWrapper = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin: 0 auto;
+  width: 100%;
+  max-width: 1400px;
+  width: 100%;
+  margin-top: 50px;
+  position: fixed;
+  justify-self: center;
+`;
