+++ src/components/templates/Error/ErrorPage.styles.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const ErrorPageWrapper = styled.div`
+  color: ${({ theme }) => theme.colors.white};
+  font-size: ${({ theme }) => theme.fontSizes.l};
+  width: 100%;
+  max-width: 1160px;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+`;
