+++ src/components/templates/Events/EventPage.styles.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const StyledEventPageWrapper = styled.div`
+  width: 100%;
+  max-width: 1160px;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+`;
+
+export default StyledEventPageWrapper;
