+++ src/components/templates/Home/HomePage.styles.tsx
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+import bg from '../../../assets/spacex-bg.png';
+
+export const HomePageContainer = styled.div`
+  width: 100%;
+  max-width: 1160px;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+`;
+
+export const HomePageBody = styled.div`
+  max-width: 680px;
+`;
