--- src/views/App.test.tsx
@@ -1,8 +1,0 @@
-import { render } from '@testing-library/react';
-import App from './App';
-
-describe('App', () => {
-  test('renders App component', () => {
-    render(<App />);
-  });
-});
