--- src/views/App.tsx
@@ -1,16 +1,0 @@
-import { FC } from 'react';
-import GlobalStyle from '../theme/GlobalStyle';
-import { ThemeProvider } from 'styled-components';
-import theme from '../theme/mainTheme';
-import { Routes } from '../routes/Routes';
-
-const App: FC = () => (
-  <>
-    <GlobalStyle />
-    <ThemeProvider theme={theme}>
-      <Routes />
-    </ThemeProvider>
-  </>
-);
-
-export default App;
