--- src/views/Contact/ContactPage.styles.tsx
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-
-export const ContactPageWrapper = styled.div`
-  color: ${({ theme }) => theme.colors.white};
-  font-size: ${({ theme }) => theme.fontSizes.l};
-`;
