--- src/views/Error/ErrorPage.styles.tsx
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-
-export const ErrorPageWrapper = styled.div`
-  color: ${({ theme }) => theme.colors.white};
-  font-size: ${({ theme }) => theme.fontSizes.l};
-`;
