--- src/views/Events/EventsPage.tsx
@@ -1,6 +1,0 @@
-import EventList from '../../components/Event/EventList';
-import { FC } from 'react';
-
-export const EventsPage: FC = () => {
-  return <EventList />;
-};
