+++ src/views/Root.styles.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+const StyledRootWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  max-width: 1400px;
+  height: 100vh;
+  justify-content: center;
+  align-items: center;
+  margin: 0 auto;
+`;
+
+export default StyledRootWrapper;
