+++ src/views/Root.test.tsx
@@ -1,0 +1,8 @@
+import { render } from '@testing-library/react';
+import Root from './Root';
+
+describe('Root', () => {
+  test('renders App component', () => {
+    render(<Root />);
+  });
+});
